requirements.txt
# === Core API ===
fastapi==0.115.0
uvicorn[standard]==0.30.6

# === Database ===
sqlalchemy==2.0.34
psycopg[binary]==3.2.2
alembic==1.13.2

# === Configuration ===
pydantic-settings==2.5.2

# === Security ===
python-jose[cryptography]==3.3.0
passlib==1.7.4

# === Testing ===
pytest==8.3.3
httpx==0.27.2